Fix binnor B_pair struct and sorting
This was broken as struct/union with raw unsigned construct in struct
binnor::B_pair got removed during CMSSW review process. 'raw' was
actually used for sorting but this never showd up as the templates did
not get instantianted, i.e., binnor was not used in cmssw.

Here we remove explicit bin1/2 variables with specified bit-size and
replace them with a single unsigned packed_value and provide
hand-written bit-shifting / masking for bin-data access.